requirements.txt
flask
requests
flask-cors
gunicorn
python-dotenv